Remove emphasis in equation
In https://docs.manifold.markets/binary-markets, the sentence
> `$100 * 13% * 10% = $1.3`

is interpreted as 
> $100 13% 10% = $1.3

with 13% in italic. Because the `*` are symbols indicating start/end of italic. I don't know what system you use, so this suggest what I expect to be the safest correction, indicating as code, which ensure it won't be interpreted. If you have mathjax installed, using \($100 \times 13% \times 10% = $1.3\) maybe even better. With the downside that Github may not be able to interpret your page as it's not markdown anymore. (not sure about it; I have seen recent news about mathjax in github, but I didn't investigate a lot actually)